+++ src/main/java/de/pascalwagler/epubcheckfx/model/ExportFormat.java
@@ -3,23 +3,34 @@ package de.pascalwagler.epubcheckfx.model;
 import de.pascalwagler.epubcheckfx.ui.Translatable;
 import lombok.Getter;
 
+@Getter
 public enum ExportFormat implements Translatable {
-    HTML("html", "html", "export_format.html"),
-    MARKDOWN("markdown", "md", "export_format.markdown"),
-    MARKDOWN_LIST("markdown-list", "md", "export_format.markdown_list"),
-    PLAINTEXT("plaintext", "txt", "export_format.plaintext"),
-    EPUB_CHECKER_TXT("epub-checker-txt", "txt", "export_format.epub-checker-txt");
+    HTML("html",
+            new String[]{"html", "htm"}, "export_format.html"),
+    MARKDOWN("markdown",
+            new String[]{"md", "markdown"}, "export_format.markdown"),
+    ASCIIDOC("asciidoc",
+            new String[]{"adoc", "asciidoc", "txt"}, "export_format.asciidoc"),
+    RESTRUCTUREDTEXT("restructuredtext",
+            new String[]{"rst"}, "export_format.restructuredtext"),
+    TEXTILE("textile",
+            new String[]{"textile"}, "export_format.textile"),
+    PLAINTEXT("plaintext",
+            new String[]{"txt"}, "export_format.plaintext"),
+    EPUB_CHECKER_TXT("epub-checker-txt",
+            new String[]{"txt"}, "export_format.epub-checker-txt"),
+    CSV("comma-separated-values",
+            new String[]{"csv"}, "export_format.csv"),
+    TSV("tab-separated-values",
+            new String[]{"tsv", "tab"}, "export_format.tsv");
 
-    ExportFormat(String name, String extension, String i18nKey) {
+    ExportFormat(String name, String[] extensions, String i18nKey) {
         this.name = name;
-        this.extension = extension;
+        this.extensions = extensions;
         this.i18nKey = i18nKey;
     }
 
-    @Getter
     private final String name;
-    @Getter
-    private final String extension;
-    @Getter
+    private final String[] extensions;
     private final String i18nKey;
 }