+++ src/main/java/de/pascalwagler/epubcheckfx/model/ReportMetadata.java
@@ -1,0 +1,10 @@
+package de.pascalwagler.epubcheckfx.model;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class ReportMetadata {
+    private String language;
+}