+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project are documented in this file. On the [releases page](https://github.com/Wandmalfarbe/EPUBCheckFX/releases) you can see all released versions of EPUBCheckFX and download the [latest version](https://github.com/Wandmalfarbe/EPUBCheckFX/releases/latest).
 
+## In Development
+
+- Add the following export formats:
+  - AsciiDoc
+  - reStructuredText
+  - Textile
+  - CSV (Comma-separated values)
+  - TSV (Tab-separated values)
+- Remove the export format *Markdown (Table)*.
+- Remove the additional extracted features from all export formats.
+- Don't show empty line numbers when exporting to *plaintext* and *pagina EPUB-Checker Text*.
+- Don't HTML encode anything when exporting to *plaintext* and *pagina EPUB-Checker Text*.
+- Reports are fully translated.
+
 ## [1.1.0] - 2023-09-04
 
 - Add a toolbar with a new button to open folders.